% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_dark.R
\name{pmf_dark_available}
\alias{pmf_dark_available}
\title{Check if Python module pmf_dark and its dependencies are available}
\usage{
pmf_dark_available()
}
\value{
`TRUE` if Python >= 3.13 is available, and both `torch` and `pmf_dark`
  are installed; `FALSE` otherwise.
}
\description{
Check if Python module pmf_dark and its dependencies are available
}
