% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_dark.R
\name{pmf_dark_module}
\alias{pmf_dark_module}
\title{Import the Python pmf_dark module}
\usage{
pmf_dark_module(delay_load = TRUE)
}
\arguments{
\item{delay_load}{Whether to delay module loading until first use. Can be a
logical value or a list of callbacks for `reticulate::import()`.}
}
\value{
A Python module object from `reticulate::import()`.
}
\description{
Import the Python pmf_dark module
}
