% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_dark.R
\name{pmf_distribution}
\alias{pmf_distribution}
\title{Predict current species distribution}
\usage{
pmf_distribution(model, pred_batch_size = NULL, return_means = TRUE)
}
\arguments{
\item{model}{A fitted `PMFDark` model object returned by `pmf_fit()`.}

\item{pred_batch_size}{Chunk size to process sites during prediction (default: NULL).}

\item{return_means}{Returns a data frame of posterior means if TRUE, or a raw array of posterior samples if FALSE.}
}
\value{
A data frame of posterior means (if TRUE) or a NumPy array of raw posterior samples (if FALSE).
}
\description{
Generates species occurrence probabilities (with latent factors) from a fitted `PMFDark` model.
}
