% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_dark.R
\name{use_python}
\alias{use_python}
\title{Configure Python Version}
\usage{
use_python(python, required = TRUE)
}
\arguments{
\item{python}{Path to the Python executable.}

\item{required}{Logical. Is this Python version required? (default: TRUE)}
}
\description{
Specifies the path to the Python binary that should be used by `reticulate`.
}
